+++ src/ui/include/horizon/ui/IconGenerator.h
@@ -133,6 +133,7 @@ private:
     static QIcon drawPatternLinear(int s);
     static QIcon drawPatternCircular(int s);
     static QIcon drawMirror3d(int s);
+    static QIcon drawHole(int s);
 };
 
 }  // namespace hz::ui